+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Support for escape and unicode characters [#98](https://github.com/ufirstgroup/ymlr/pull/98)
+- Refactor `Ymlr.Encode` to make it faster.
 
 <!--------------------- Don't add new entries after this line --------------------->